+++ QGLViewer/manipulatedCameraFrame.h
@@ -103,7 +103,7 @@ public:
 
 private:
 #ifndef DOXYGEN
-	void zoom(float delta, const Camera * const camera);
+	void zoom(qreal delta, const Camera * const camera);
 #endif
 	//@}
 
@@ -114,7 +114,7 @@ public Q_SLOTS:
 
 	Default value is 0.0, but it is modified according to the QGLViewer::sceneRadius() when the
 	ManipulatedCameraFrame is set as the Camera::frame(). */
-	void setFlySpeed(float speed) { flySpeed_ = speed; }
+	void setFlySpeed(qreal speed) { flySpeed_ = speed; }
 
 	/*! Sets the sceneUpVector(), defined in the world coordinate system.
 
@@ -131,7 +131,7 @@ public:
 
 	\attention When the ManipulatedCameraFrame is set as the Camera::frame(), this value is set
 	according to the QGLViewer::sceneRadius() by QGLViewer::setSceneRadius(). */
-	float flySpeed() const { return flySpeed_; }
+	qreal flySpeed() const { return flySpeed_; }
 
 	/*! Returns the up vector of the scene, expressed in the world coordinate system.
 
@@ -191,8 +191,8 @@ private:
 
 private:
 	// Fly mode data
-	float flySpeed_;
-	float driveSpeed_;
+	qreal flySpeed_;
+	qreal driveSpeed_;
 	Vec sceneUpVector_;
 	QTimer flyTimer_;
